requirements.txt
Flask
selenium
beautifulsoup4
psycopg2-binary
webdriver-manager==4.0.2
